test: classify the release archive guard as test code
